Backend/app/requirements.txt
fastapi
uvicorn
python-multipart
python-docx
PyMuPDF
